+++ OpenMarket/OpenMarket/Model/Images.swift
@@ -1,0 +1,8 @@
+//
+//  Images.swift
+//  OpenMarket
+//
+//  Created by 박도원 on 2022/10/26.
+//
+
+import Foundation
