+++ OpenMarket/OpenMarket/Model/Vendors.swift
@@ -1,0 +1,8 @@
+//
+//  Vendors.swift
+//  OpenMarket
+//
+//  Created by 박도원 on 2022/10/26.
+//
+
+import Foundation
