+++ OpenMarket/OpenMarket/View/Assets.xcassets/Mock.dataset/Contents.json
@@ -1,7 +1,7 @@
 {
   "data" : [
     {
-      "filename" : "products.json",
+      "filename" : "Mock.json",
       "idiom" : "universal"
     }
   ],
